+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Query page: added default query after searching for first dataset
 ### Changed
 - Readme section: modified 'Run' button into a link
 - Hint extension for tile widget account